+++ spec/langchain/vectorsearch/pinecone_spec.rb
@@ -5,24 +5,32 @@ require "pinecone"
 RSpec.describe Langchain::Vectorsearch::Pinecone do
   let(:index_name) { "documents" }
   let(:namespace) { "namespaced" }
+  let(:llm) { Langchain::LLM::OpenAI.new(api_key: "123") }
 
   subject {
     described_class.new(
       environment: "test",
       api_key: "secret",
       index_name: index_name,
-      llm: Langchain::LLM::OpenAI.new(api_key: "123")
+      llm: llm
     )
   }
 
   describe "#create_default_schema" do
     it "returns true" do
-      allow_any_instance_of(Pinecone::Client).to receive(:create_index).with(
+      allow(subject.client).to receive(:create_index).with(
         metric: described_class::DEFAULT_METRIC,
         name: index_name,
         dimension: subject.default_dimension
       ).and_return(true)
       expect(subject.create_default_schema).to eq(true)
+    end
+  end
+
+  describe "#destroy_default_schema" do
+    it "returns true" do
+      allow(subject.client).to receive(:delete_index).with(index_name).and_return(true)
+      expect(subject.destroy_default_schema).to eq(true)
     end
   end
 
@@ -132,6 +140,53 @@ RSpec.describe Langchain::Vectorsearch::Pinecone do
       it "adds texts" do
         expect(subject.add_texts(texts: [text], metadata: metadata)).to eq(true)
       end
+    end
+
+    describe "with ids" do
+      let!(:vectors) do
+        [
+          {
+            id: "456",
+            metadata: {content: text},
+            values: embedding
+          }
+        ]
+      end
+
+      before(:each) do
+        allow_any_instance_of(Pinecone::Index).to receive(:upsert).with(
+          vectors: vectors, namespace: ""
+        ).and_return(true)
+      end
+
+      it "adds texts" do
+        expect(subject.add_texts(texts: [text], ids: [456])).to eq(true)
+      end
+    end
+  end
+
+  describe "#update_texts" do
+    let(:vectors) do
+      [
+        {
+          id: "123",
+          metadata: {content: text},
+          values: embedding
+        }
+      ]
+    end
+
+    before do
+      vector = double(Pinecone::Vector)
+      allow(subject.llm).to receive(:embed).with(text: text).and_return(embedding)
+      allow(subject.client).to receive(:index).with(index_name).and_return(vector)
+      allow(vector).to receive(:update).with(
+        values: embedding, id: "123", namespace: "", set_metadata: nil
+      ).and_return(true)
+    end
+
+    it "updates texts" do
+      expect(subject.update_texts(texts: [text], ids: [123])).to eq([true])
     end
   end