+++ spec/langchain/vectorsearch/weaviate_spec.rb
@@ -16,7 +16,7 @@ RSpec.describe Langchain::Vectorsearch::Weaviate do
     let(:fixture) { JSON.parse(File.read("spec/fixtures/vectorsearch/weaviate_create_default_schema.json")) }
 
     before do
-      allow_any_instance_of(Weaviate::Client).to receive_message_chain(:schema, :create).and_return(fixture)
+      allow(subject.client).to receive_message_chain(:schema, :create).and_return(fixture)
     end
 
     it "creates the default schema" do
@@ -24,39 +24,81 @@ RSpec.describe Langchain::Vectorsearch::Weaviate do
     end
   end
 
+  describe "#create_default_schema" do
+    before do
+      allow(subject.client).to receive_message_chain(:schema, :delete).and_return(true)
+    end
+
+    it "creates the default schema" do
+      expect(subject.destroy_default_schema).to eq(true)
+    end
+  end
+
   describe "#add_texts" do
     let(:fixture) { JSON.parse(File.read("spec/fixtures/vectorsearch/weaviate_add_texts.json")) }
 
-    before do
-      allow_any_instance_of(
-        Weaviate::Objects
+    def stub(id)
+      allow(
+        subject.client.objects
       ).to receive(:batch_create)
         .with(
           objects: [{
             class: "products",
-            properties: {content: "Hello World"},
+            properties: {__id: id.to_s, content: "Hello World"},
             vector: [-0.0018150936, 0.0017554426, -0.022715086]
           }]
         )
         .and_return(fixture)
 
-      allow_any_instance_of(
-        ::OpenAI::Client
-      ).to receive(:embeddings).and_return({
-        "data" => [
-          {
-            "embedding" => [
-              -0.0018150936,
-              0.0017554426,
-              -0.022715086
-            ]
-          }
-        ]
-      })
+      allow(subject.llm).to receive(:embed).and_return([
+        -0.0018150936,
+        0.0017554426,
+        -0.022715086
+      ])
     end
 
-    it "adds texts" do
-      expect(subject.add_texts(texts: ["Hello World"])).to eq(fixture)
+    context "with ids" do
+      before do
+        stub(1)
+      end
+
+      it "adds texts" do
+        expect(subject.add_texts(texts: ["Hello World"], ids: [1])).to eq(fixture)
+      end
+    end
+
+    context "without ids" do
+      before do
+        stub(nil)
+      end
+
+      it "adds texts" do
+        expect(subject.add_texts(texts: ["Hello World"])).to eq(fixture)
+      end
+    end
+  end
+
+  describe "#update_texts" do
+    let(:fixture) { JSON.parse(File.read("spec/fixtures/vectorsearch/weaviate_add_texts.json")) }
+
+    let(:record) {
+      [{"_additional" => {"id" => "372ba500-01af-4448-aa03-21f3dd25a456"}}]
+    }
+
+    before do
+      allow(subject.client.query).to receive(:get).and_return(record)
+
+      allow(subject.llm).to receive(:embed).and_return([
+        -0.0018150936,
+        0.0017554426,
+        -0.022715086
+      ])
+
+      allow(subject.client.objects).to receive(:update).and_return(fixture.first)
+    end
+
+    it "updates texts" do
+      expect(subject.update_texts(texts: ["Hello World"], ids: [1])).to eq(fixture)
     end
   end
 
@@ -64,30 +106,22 @@ RSpec.describe Langchain::Vectorsearch::Weaviate do
     let(:fixture) { JSON.parse(File.read("spec/fixtures/vectorsearch/weaviate_search.json")) }
 
     before do
-      allow_any_instance_of(
-        Weaviate::Query
+      allow(
+        subject.client.query
       ).to receive(:get)
         .with(
           class_name: "products",
           near_vector: "{ vector: [-0.0018150936, 0.0017554426, -0.022715086] }",
           limit: "4",
-          fields: "content _additional { id }"
+          fields: "__id content _additional { id }"
         )
         .and_return(fixture)
 
-      allow_any_instance_of(
-        ::OpenAI::Client
-      ).to receive(:embeddings).and_return({
-        "data" => [
-          {
-            "embedding" => [
-              -0.0018150936,
-              0.0017554426,
-              -0.022715086
-            ]
-          }
-        ]
-      })
+      allow(subject.llm).to receive(:embed).and_return([
+        -0.0018150936,
+        0.0017554426,
+        -0.022715086
+      ])
     end
 
     it "searches for similar texts" do
@@ -99,14 +133,14 @@ RSpec.describe Langchain::Vectorsearch::Weaviate do
     let(:fixture) { JSON.parse(File.read("spec/fixtures/vectorsearch/weaviate_search.json")) }
 
     before do
-      allow_any_instance_of(
-        Weaviate::Query
+      allow(
+        subject.client.query
       ).to receive(:get)
         .with(
           class_name: "products",
           near_vector: "{ vector: [0.1, 0.2, 0.3] }",
           limit: "4",
-          fields: "content _additional { id }"
+          fields: "__id content _additional { id }"
         )
         .and_return(fixture)
     end